+++ modules/visual-tests/src/routes/test/MainMenu.tsx
@@ -35,6 +35,10 @@ export const mainMenuItems: MainMenuItemList = [
     label: 'Mannlíf',
     href: '#MegaMenu:Mannlíf',
   },
+  {
+    label: 'Link test',
+    href: '?blu=blu',
+  },
   '---',
   {
     label: 'EN',